module3-task1 add data.php and functions.php
