+++ data.php
@@ -1,0 +1,40 @@
+<?php
+$projects = ["Входящие", "Учеба", "Работа", "Домашние дела", "Авто"];
+$tasks = [
+	[
+		'title' => 'Собеседование в IT компании',
+		'date_completed' => '01.12.2019',
+		'category' => 'Работа',
+		'executed' => false
+	],
+	[
+		'title' => 'Выполнить тестовое задание',
+		'date_completed' => '25.12.2019',
+		'category' => 'Работа',
+		'executed' => false
+    ],
+    [
+		'title' => 'Сделать задание первого раздела',
+		'date_completed' => '21.12.2019',
+		'category' => 'Учеба',
+		'executed' => true
+    ],
+    [
+		'title' => 'Встреча с другом',
+		'date_completed' => '22.12.2019',
+		'category' => 'Входящие',
+		'executed' => false
+    ],
+    [
+		'title' => 'Купить корм для кота',
+		'date_completed' => null,
+		'category' => 'Домашние дела',
+		'executed' => false
+    ],
+    [
+		'title' => 'Заказать пиццу',
+		'date_completed' => null,
+		'category' => 'Домашние дела',
+		'executed' => false
+    ]
+];
